Refactor de_export.py, extract sympy_utils.py
No changes in functionality.

Related to #2306.
